requirements.txt
amqp==5.0.6
asgiref==3.2.10
boto3==1.17.58
botocore==1.20.112
celery==5.1.1
channels==3.0.0
channels-redis==3.4.0
daphne==3.0.2
Django==3.1.2
django-cors-headers==3.7.0
django-storages==1.11.1
django-tenants==3.3.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
flower==1.0.0
gunicorn==19.6.0
Pillow==8.2.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
pytest==7.1.2
factory-boy==3.2.1
python-dotenv==0.18.0